Implement dummy tl_detector
Tl detector now finds closest traffic light ahead of the car and publishes a traffic_waupoint message with the location of the light stop line and the current state of the light. Light state is taken from the ground truth message stream. This will have to be replaced with a call to tl_classifier when it's ready.